Add files via upload
This project was focused  on creating a predictive model for categorization of reviews from customers into 'positive' or 'negative' feedback. I have applied LSTM model (a type of RNN model) of Deep Learning Classification algorithm on reviews and feedback dataset of 50,000 customers of a company. The dataset was preprocessed and the model was successfully trained and deployed on web for testing the random reviews. Model showed an overall accuracy of approx. 70%